Move web builder defaults helper to app bundle
